+++ src/helpers/parseEngagementUrl.ts
@@ -1,6 +1,12 @@
-import { isEngagementType } from '@/helpers/isEngagementType.js';
 import { isSocialSource } from '@/helpers/isSocialSource.js';
 import { resolveSourceFromUrlNoFallback } from '@/helpers/resolveSource.js';
+import { EngagementType } from '@/constants/enum.js';
+
+function isEngagementType(value: string): value is EngagementType {
+    return [EngagementType.Mirrors, EngagementType.Quotes, EngagementType.Likes, EngagementType.Recasts].includes(
+        value as EngagementType,
+    );
+}
 
 export function parseOldEngagementUrl(url: URL) {
     if (!url.pathname.startsWith('/post')) return null;
