+++ src/helpers/parseProfileUrl.ts
@@ -4,6 +4,7 @@ import { isSocialProfileCategory } from '@/helpers/isSocialProfileCategory.js';
 import { isSocialSource } from '@/helpers/isSocialSource.js';
 import { isWalletProfileCategory } from '@/helpers/isWalletProfileCategory.js';
 import { resolveSourceFromUrlNoFallback } from '@/helpers/resolveSource.js';
+import { isProfilePageSource } from '@/helpers/isProfilePageSource.js';
 
 export function parseProfileUrl(pathname: string) {
     if (!pathname.startsWith('/profile')) return null;
@@ -23,4 +24,14 @@ export function parseProfileUrl(pathname: string) {
             : null;
     }
     return null;
+}
+
+export function parseOldProfileUrl(url: URL) {
+    if (!url.pathname.startsWith('/profile')) return null;
+    const source = resolveSourceFromUrlNoFallback(url.searchParams.get('source'));
+    if (!source || !isProfilePageSource(source)) return null;
+    const [, , id, ...end] = url.pathname.split('/');
+    if (end.length) return null;
+    if (!id) return null;
+    return { source, id };
 }
